% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congGenerator.R
\name{congGenerator}
\alias{congGenerator}
\title{Linear Congruential Generator}
\usage{
congGenerator(n, a, c = 0, m, seed)
}
\arguments{
\item{n}{Integer. Sequence length.}

\item{a}{Integer. The multiplier used in the generator.}

\item{c}{Integer, The additive constant. Default is 0.}

\item{m}{Integer. The modulus used in the generator.}

\item{seed}{Integer. Initial value that determines the starting point of the sequence.}
}
\value{
A numeric vector of length \code{n} containing the sequence of the uniform pseudo-random numbers.
}
\description{
Implements a linear congruential generator to create a sequence of pseudo-random numbers from an uniform distribution U(0, 1).
}
\details{
The linear congruential generator is one of the oldest and best-known algorithms for generating pseudo-random numbers.
The quality of the generated sequence depends on the appropriate choice of parameters \code{a}, \code{c}, and \code{m}.

The formula used is: \deqn{x_{i+1} = (a x_i + c) \mod m}
}
\examples{
# Generate a sequence with arbitrary parameters(bad choose)
congGenerator(n = 100, seed = 12, a = 64, c = 11, m = 7)

# Generate another sequence without an additive constant
congGenerator(n = 100, seed = 2003, a = 45, m = 2^7)

# Evaluate the quality of a sequence
congGenerator(n = 100, seed = 1610, a = 21, c = 25, m = 2^10)


}
