% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom.R
\name{geom}
\alias{geom}
\title{Geometric random variable generator}
\usage{
geom(n, p)
}
\arguments{
\item{n}{Number of random values to generate.}

\item{p}{Probability of success in each trial (0 < p < 1).}
}
\value{
A numeric vector of length n containing the number of trials required
to obtain the first success in each simulation.
}
\description{
This function generates random numbers following a geometric distribution
using direct simulation. For each observation, independent trials are repeated
until the first success occurs, counting the number of trials required.
}
\examples{
# Generate 10 values with success probability 0.3
# geom(10,0.3)
}
\references{
Ross, S. M. (2014). Introduction to Probability Models.
Devore, J. L. (2012). Probability and Statistics for Engineering and the Sciences.
}
