% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleSquare.R
\name{middleSquare}
\alias{middleSquare}
\title{This function implements the middle-square method to generate pseudo-random numbers in the interval (0,1)}
\usage{
middleSquare(n, seed)
}
\arguments{
\item{n}{length of the generated sequence}

\item{seed}{The initial seed used to start the algorithm}
}
\value{
Returns a vector of pseudo-random numbers generated by the middle-square method
}
\description{
This function implements the middle-square method to generate pseudo-random numbers in the interval (0,1)
}
\details{
This function applies the middle-square algorithm, which consists of squaring the current seed and extracting the central digits to form the next seed. Let \eqn{X_0} be the initial seed with \eqn{d} digits. At each iteration, \eqn{X_i^2} is computed and the middle \eqn{d} digits are extracted to define \eqn{X_{i+1}}. The generated pseudo-random number is obtained as \eqn{U_i = X_i / 10^d}. This process is repeated \eqn{n} times.
}
\examples{
middleSquare(10, 5497)

}
