% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period}
\alias{period}
\title{This function detects the period for a random sequence}
\usage{
period(x)
}
\arguments{
\item{x}{The sequence for detecting the period}
}
\value{
Returns the period of a sequence
}
\description{
This function detects the period for a random sequence
}
\examples{
sample <- runif(100)
period(sample)

}
