% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifTestChisq.R
\name{unifTestChisq}
\alias{unifTestChisq}
\title{This function implements the goodness of fit chi-square test to prove if a sample comes from a uniform distribution U(0, 1)}
\usage{
unifTestChisq(x)
}
\arguments{
\item{x}{The sample to be tested}
}
\value{
Returns the chi square statistic, the degree of freedom and the p-value for the goodness of fit test
}
\description{
This function implements the goodness of fit chi-square test to prove if a sample comes from a uniform distribution U(0, 1)
}
\details{
This function contrast the hypotheses \eqn{H_0: \pi = \hat{\pi}} Vs. \eqn{H_1: \pi \neq \hat{\pi}}. To facilitate the compute, we always use a partition of 10 intervals to compute the frequencies in the sample, as to compute the theoretical probabilities.
}
\examples{
sample <- runif(100)
unifTestChisq(sample)

}
