% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifTestKS.R
\name{unifTestKS}
\alias{unifTestKS}
\title{This function implements the goodness of fit Kolmogorov-Smirnov test to prove if a sample comes from a uniform distribution U(0, 1)}
\usage{
unifTestKS(x)
}
\arguments{
\item{x}{The sample to be tested}
}
\value{
Returns the D statistic and the p-value for the goodness of fit test
}
\description{
This function implements the goodness of fit Kolmogorov-Smirnov test to prove if a sample comes from a uniform distribution U(0, 1)
}
\details{
This function contrasts the hypotheses \eqn{H_0: F_n(x) = F(x)} vs \eqn{H_0: F_n(x) \neq F(x)}. Where \eqn{F_n(x)} is the empirical cumulate function distribution and \eqn{F(x)} is the theoretical uniform cumulate distribution function.
}
\examples{
sample <- runif(100)
unifTestKS(sample)

}
